requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
runpod
Pillow